solutionAgent/requirements.txt
langchain
langchainhub
openai
langchain-community
tavily-python
python-dotenv
uvicorn
huggingface-hub==0.20.3
fastapi
lxml_html_clean
newspaper3k